core: migrate `AppRouter` to new location
